为Controller中的方法添加了注释

+++ theThird/src/main/java/com/example/cinema/bl/promotion/CouponService.java
@@ -18,7 +18,7 @@ public interface CouponService {
 
     ResponseVO getSpecialCoupon();
 
-    ResponseVO updateSpecialCouponDiscount(int discount);
+    ResponseVO updateSpecialCouponDiscount(double discount);
 
     ResponseVO getAllUserByConsume(int consume);
 }
